+++ src/galapix/layouter/regular_layouter.cpp
@@ -33,9 +33,9 @@ RegularLayouter::layout(const ImageCollection& images)
 {
   if (!images.empty())
   {
-    int w = int(Math::sqrt(m_aspect_w * static_cast<float>(images.size()) / m_aspect_h));
+    size_t w = static_cast<size_t>(Math::sqrt(m_aspect_w * static_cast<float>(images.size()) / m_aspect_h));
 
-    for(int i = 0; i < int(images.size()); ++i)
+    for(size_t i = 0; i < images.size(); ++i)
     {
       float target_scale = Math::min(1000.0f / static_cast<float>(images[i]->get_original_width()),
                                      1000.0f / static_cast<float>(images[i]->get_original_height()));
@@ -45,12 +45,12 @@ RegularLayouter::layout(const ImageCollection& images)
       if ((i/w) % 2 == 0)
       {
         images[i]->set_pos(Vector2f(static_cast<float>(i % w) * 1024.0f,
-                                           static_cast<float>(i / w) * 1024.0f));
+                                    static_cast<float>(i / w) * 1024.0f));
       }
       else
       {
         images[i]->set_pos(Vector2f(static_cast<float>(w - (i % w)-1) * 1024.0f,
-                                           static_cast<float>(i / w)         * 1024.0f));
+                                    static_cast<float>(i / w) * 1024.0f));
       }
     }
   }
