+++ src/util/sha1.cpp
@@ -117,7 +117,7 @@ SHA1::from_string(const std::string& str)
   else
   {
     uint8_t data[20];
-    for(int i = 0; i < 20; ++i)
+    for(size_t i = 0; i < 20; ++i)
     {
       data[i] = static_cast<uint8_t>(hex2int(str[2*i]) * 0x10 + hex2int(str[2*i+1]));
     }
